+++ DynatracePS/Public/Get-DynatraceEntityType.ps1
@@ -6,11 +6,17 @@ function Get-DynatraceEntityType {
     .DESCRIPTION
         Get list of entity types in Dynatrace environment
 
+    .PARAMETER Type
+        Specific entity type to get properties for
+
     .PARAMETER OutputAsJson
         Output the properties as a JSON string
 
     .EXAMPLE
         Get-DynatraceEntityType
+
+    .EXAMPLE
+        Get-DynatraceEntityType -Type SERVICE
 
     .NOTES
         https://api.dynatrace.com/spec/#/
@@ -18,6 +24,7 @@ function Get-DynatraceEntityType {
 
     [CmdletBinding()]
     param(
+        [string]$Type,
         [switch]$OutputAsJson
     )
 
@@ -33,14 +40,23 @@ function Get-DynatraceEntityType {
             pageSize = 200
         }
 
+        if ($Type) {
+            $uri = "https://$EnvironmentID.live.dynatrace.com/api/v2/entityTypes/$Type"
+            $splatParameters = @{
+                Uri = $uri
+            }
+        } else {
+            $splatParameters = @{
+                Uri = $uri
+                GetParameter = $GetParameter
+                RestResponseProperty = 'types'
+            }    
+        }
+
         Write-Verbose "[$($MyInvocation.MyCommand.Name)] $Uri"
     }
     process {
-        $splatParameters = @{
-            Uri = $uri
-            GetParameter = $GetParameter
-            RestResponseProperty = 'types'
-        }
+
         $output = Invoke-DynatraceAPIMethod @splatParameters
         if ($OutputAsJson) {
             $output | ConvertTo-Json -Depth 6